requirements.txt
Django==4.1.4
psycopg2==2.9.5
python-dotenv==0.21.0
whitenoise==6.2.0
werkzeug==2.2.2